🐛 Fix win crash

+++ windows/internal/ZegoExpressEngineMethodHandler.cpp
@@ -1025,7 +1025,7 @@ void ZegoExpressEngineMethodHandler::enableVideoObjectSegmentation(
             EXPRESS::ZegoBackgroundConfig backgroundConfig;
             backgroundConfig.processType = (EXPRESS::ZegoBackgroundProcessType)std::get<int32_t>(backgroundConfigMap[FTValue("processType")]);
             backgroundConfig.blurLevel = (EXPRESS::ZegoBackgroundBlurLevel)std::get<int32_t>(backgroundConfigMap[FTValue("blurLevel")]);
-            backgroundConfig.color = std::get<int32_t>(backgroundConfigMap[FTValue("color")]);
+            backgroundConfig.color = std::get<int64_t>(backgroundConfigMap[FTValue("color")]);
             backgroundConfig.imageURL = std::get<std::string>(backgroundConfigMap[FTValue("imageURL")]);
             backgroundConfig.videoURL = std::get<std::string>(backgroundConfigMap[FTValue("videoURL")]);
